Fixed a minor bug with float32 and float64 variable

+++ physics/models/models.go
@@ -7,7 +7,7 @@ type Simulation struct {
 	Time      *int
 	Last_time float64
 
-	E float32
+	E float64
 }
 
 type Object struct {
